+++ Apps/CesiumViewer/CesiumViewer.js
@@ -10,7 +10,7 @@ define([
         'Cesium/DataSources/CzmlDataSource',
         'Cesium/DataSources/GeoJsonDataSource',
         'Cesium/DataSources/KmlDataSource',
-        'Cesium/Scene/TileMapServiceImageryProvider',
+        'Cesium/Scene/createTileMapServiceImageryProvider',
         'Cesium/Widgets/Viewer/Viewer',
         'Cesium/Widgets/Viewer/viewerCesiumInspectorMixin',
         'Cesium/Widgets/Viewer/viewerDragDropMixin',
@@ -26,7 +26,7 @@ define([
         CzmlDataSource,
         GeoJsonDataSource,
         KmlDataSource,
-        TileMapServiceImageryProvider,
+        createTileMapServiceImageryProvider,
         Viewer,
         viewerCesiumInspectorMixin,
         viewerDragDropMixin) {
@@ -47,7 +47,7 @@ define([
 
     var imageryProvider;
     if (endUserOptions.tmsImageryUrl) {
-        imageryProvider = new TileMapServiceImageryProvider({
+        imageryProvider = createTileMapServiceImageryProvider({
             url : endUserOptions.tmsImageryUrl
         });
     }