+++ Source/Scene/GoogleEarthImageryProvider.js
@@ -84,7 +84,7 @@ define([
      * @see BingMapsImageryProvider
      * @see createOpenStreetMapImageryProvider
      * @see SingleTileImageryProvider
-     * @see TileMapServiceImageryProvider
+     * @see createTileMapServiceImageryProvider
      * @see WebMapServiceImageryProvider
      * @see WebMapTileServiceImageryProvider
      * @see UrlTemplateImageryProvider
@@ -591,4 +591,3 @@ define([
 
     return GoogleEarthImageryProvider;
 });
-