+++ Source/Scene/ImageryProvider.js
@@ -233,7 +233,8 @@ define([
          * include an alpha channel.  If this property is false, an alpha channel, if present, will
          * be ignored.  If this property is true, any images without an alpha channel will be treated
          * as if their alpha is 1.0 everywhere.  When this property is false, memory usage
-         * and texture upload time are reduced.
+         * and texture upload time are reduced.  This function should
+         * not be called before {@link ImageryProvider#ready} returns true.
          * @memberof ImageryProvider.prototype
          * @type {Boolean}
          * @readonly