+++ Source/Scene/OpenStreetMapImageryProvider.js
@@ -51,7 +51,7 @@ define([
      * @see BingMapsImageryProvider
      * @see GoogleEarthImageryProvider
      * @see SingleTileImageryProvider
-     * @see TileMapServiceImageryProvider
+     * @see createTileMapServiceImageryProvider
      * @see WebMapServiceImageryProvider
      * @see WebMapTileServiceImageryProvider
      * @see UrlTemplateImageryProvider
@@ -64,7 +64,7 @@ define([
      * var osm = new Cesium.OpenStreetMapImageryProvider({
      *     url : '//a.tile.openstreetmap.org/'
      * });
-     * 
+     *
      * @deprecated
      */
     function OpenStreetMapImageryProvider(options) {