+++ Source/Scene/WebMapTileServiceImageryProvider.js
@@ -65,7 +65,7 @@ define([
      * @see GoogleEarthImageryProvider
      * @see createOpenStreetMapImageryProvider
      * @see SingleTileImageryProvider
-     * @see TileMapServiceImageryProvider
+     * @see createTileMapServiceImageryProvider
      * @see WebMapServiceImageryProvider
      * @see UrlTemplateImageryProvider
      *